feat: task based camera
